[Refactor] Fix circular dependencies between trashspot and suggestion modules
This commit resolves the circular dependency issue between the trashspot and suggestion modules,
enabling the ModulithArchitectureTest to pass.

Changes:
1. Move TrashType enum from trashspot to common module
   - Updated all imports across 35 files to use com.sseudam.common.TrashType

2. Separate TrashSpot creation logic using event-driven architecture
   - Create TrashSpotSuggestionEventHandler in trashspot module
   - Move TrashSpot creation logic from SuggestionUpdateHandler to TrashSpotSuggestionEventHandler
   - Now trashspot module listens to SuggestionUpdateEvent and creates TrashSpot accordingly

3. Remove TrashSpotService dependency from suggestion module
   - Replace TrashSpotService with direct usage of TrashSpotReader and TrashSpotValidator in SuggestionFacade
   - Implement validation logic directly in SuggestionFacade

Dependency structure after fix:
- suggestion → trashspot.component (one-way, acceptable)
- trashspot.event → suggestion.event (one-way, acceptable)
- No circular dependencies

Related: Spring Modulith architecture test
