+++ sseudam-admin/src/main/kotlin/com/sseudam/admin/application/report/AdminSpotReportDetail.kt
@@ -6,7 +6,7 @@ import com.sseudam.common.Region
 import com.sseudam.report.ReportStatus
 import com.sseudam.report.ReportType
 import com.sseudam.report.SpotReport
-import com.sseudam.trashspot.TrashType
+import com.sseudam.common.TrashType
 import com.sseudam.user.UserProfile
 import java.time.LocalDateTime
