+++ sseudam-admin/src/main/kotlin/com/sseudam/admin/presentation/response/report/SpotReportAdminResponse.kt
@@ -5,7 +5,7 @@ import com.sseudam.common.Address
 import com.sseudam.common.GeoJson
 import com.sseudam.report.ReportStatus
 import com.sseudam.report.ReportType
-import com.sseudam.trashspot.TrashType
+import com.sseudam.common.TrashType
 import io.swagger.v3.oas.annotations.media.Schema
 import java.time.LocalDateTime
