+++ sseudam-core/core-api/src/main/kotlin/com/sseudam/presentation/v1/history/response/HistoryResponse.kt
@@ -5,7 +5,7 @@ import com.sseudam.common.GeoJson
 import com.sseudam.history.dto.HistoryStatus
 import com.sseudam.history.dto.SpotActionType
 import com.sseudam.history.dto.SpotHistory
-import com.sseudam.trashspot.TrashType
+import com.sseudam.common.TrashType
 import io.swagger.v3.oas.annotations.media.Schema
 import java.time.LocalDateTime
