+++ sseudam-core/core-api/src/main/kotlin/com/sseudam/presentation/v1/report/request/SpotReportCreateRequest.kt
@@ -3,7 +3,7 @@ package com.sseudam.presentation.v1.report.request
 import com.sseudam.common.Region
 import com.sseudam.report.ReportType
 import com.sseudam.report.SpotReport
-import com.sseudam.trashspot.TrashType
+import com.sseudam.common.TrashType
 import io.swagger.v3.oas.annotations.media.Schema
 
 @Schema(description = "신고하기 요청 Json")
