+++ sseudam-core/core-api/src/main/kotlin/com/sseudam/presentation/v1/report/response/SpotReportResponse.kt
@@ -6,7 +6,7 @@ import com.sseudam.common.Region
 import com.sseudam.report.ReportStatus
 import com.sseudam.report.ReportType
 import com.sseudam.report.SpotReport
-import com.sseudam.trashspot.TrashType
+import com.sseudam.common.TrashType
 import io.swagger.v3.oas.annotations.media.Schema
 import java.time.LocalDateTime
