+++ sseudam-core/core-api/src/main/kotlin/com/sseudam/presentation/v1/suggestion/response/SpotSuggestionDetailResponse.kt
@@ -5,7 +5,7 @@ import com.sseudam.common.GeoJson
 import com.sseudam.common.Region
 import com.sseudam.suggestion.SpotSuggestion
 import com.sseudam.suggestion.SuggestionStatus
-import com.sseudam.trashspot.TrashType
+import com.sseudam.common.TrashType
 import io.swagger.v3.oas.annotations.media.Schema
 import java.time.LocalDateTime
