+++ sseudam-core/core-api/src/main/kotlin/com/sseudam/presentation/v1/trashspot/response/TrashSpotResponse.kt
@@ -4,7 +4,7 @@ import com.sseudam.common.Address
 import com.sseudam.common.GeoJson
 import com.sseudam.common.Region
 import com.sseudam.trashspot.TrashSpot
-import com.sseudam.trashspot.TrashType
+import com.sseudam.common.TrashType
 import io.swagger.v3.oas.annotations.media.Schema
 import java.time.LocalDateTime
