+++ sseudam-core/core-api/src/test/kotlin/com/sseudam/application/trashspot/TrashSpotServiceTest.kt
@@ -7,7 +7,7 @@ import com.sseudam.common.GeoJson
 import com.sseudam.common.Region
 import com.sseudam.trashspot.TrashSpot
 import com.sseudam.trashspot.TrashSpotService
-import com.sseudam.trashspot.TrashType
+import com.sseudam.common.TrashType
 import com.sseudam.trashspot.component.TrashSpotAppender
 import com.sseudam.trashspot.component.TrashSpotReader
 import com.sseudam.trashspot.component.TrashSpotValidator
