+++ sseudam-core/core-api/src/test/kotlin/com/sseudam/fixture/trashspot/TrashSpotFixture.kt
@@ -7,7 +7,7 @@ import com.sseudam.common.Region
 import com.sseudam.test.helper.fixtureBuilder
 import com.sseudam.test.helper.fixtureBuilders
 import com.sseudam.trashspot.TrashSpot
-import com.sseudam.trashspot.TrashType
+import com.sseudam.common.TrashType
 import com.sseudam.trashspot.image.TrashSpotImage
 import net.jqwik.api.Arbitraries
