+++ sseudam-core/core-application/src/main/kotlin/com/sseudam/trashspot/repository/TrashSpotRepository.kt
@@ -2,7 +2,7 @@ package com.sseudam.trashspot.repository
 
 import com.sseudam.common.Region
 import com.sseudam.trashspot.TrashSpot
-import com.sseudam.trashspot.TrashType
+import com.sseudam.common.TrashType
 import com.sseudam.trashspot.dto.TrashSpotLocation
 import org.locationtech.jts.geom.Point
