+++ sseudam-core/core-domain/src/main/kotlin/com/sseudam/common/TrashType.kt
@@ -1,4 +1,4 @@
-package com.sseudam.trashspot
+package com.sseudam.common
 
 enum class TrashType(
     val displayName: String,
