+++ sseudam-storage/db-core/src/main/kotlin/com/sseudam/storage/db/core/report/SpotReportEntity.kt
@@ -7,7 +7,7 @@ import com.sseudam.report.ReportStatus
 import com.sseudam.report.ReportType
 import com.sseudam.report.SpotReport
 import com.sseudam.storage.db.core.support.BaseEntity
-import com.sseudam.trashspot.TrashType
+import com.sseudam.common.TrashType
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.EnumType
