+++ sseudam-storage/db-core/src/main/kotlin/com/sseudam/storage/db/core/trashspot/TrashSpotJpaRepository.kt
@@ -2,7 +2,7 @@ package com.sseudam.storage.db.core.trashspot
 
 import com.linecorp.kotlinjdsl.support.spring.data.jpa.repository.KotlinJdslJpqlExecutor
 import com.sseudam.common.Region
-import com.sseudam.trashspot.TrashType
+import com.sseudam.common.TrashType
 import org.locationtech.jts.geom.Point
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
